requirements.txt
django==4.1.7
beautifulsoup4
django-bootstrap4
django-crispy-forms
crispy-bootstrap4
Pillow
django-countries
reportlab
numpy
django-storages
boto3



	
